+++ cfg/LidarFilters.cfg
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-PACKAGE = "lidar_filters_pkg"
+PACKAGE = "urban_road_filter"
 
 from dynamic_reconfigure.parameter_generator_catkin import *
